Update docker-toolbelt to v3.0.1 and add a basic test suite
We update docker-toolbelt to v3, which removes promisified `Async`-suffixed functions
and instead uses dockerode's promise support.

We add basic tests that rely on a Dockerfile.test for docker-in-docker,
which gives us an environment where we can create and apply deltas without polluting the host.

The tests create and apply a delta for a simple pair of images. They require being able to
run privileged containers for docker-in-docker to work.

Change-Type: patch
Signed-off-by: Pablo Carranza Velez <pablo@resin.io>