Avoid applyDelta hangs by making waitAsync work properly for processes that have already exited
When calling ps.waitAsync for a child process ps that has already exited, the 'exit' event would
not be fired again, so the promise would never resolve. This patch adds handlers for the 'exit'
and 'error' events as soon as the process is spawned, so that waitAsync can check if the process
has already exited and handle this situation properly.

This bug would cause applyDelta to occasionally hang because rsync had exited before waitAsync was called.

Change-Type: patch
Signed-off-by: Pablo Carranza Velez <pablo@resin.io>